added node class as inner in prefixTree and changed getEdges method to findChild
